Move common components to nested directory
