+++ app/src/frontend/building-view.tsx
@@ -3,9 +3,9 @@ import { Link, NavLink } from 'react-router-dom';
 import PropTypes from 'prop-types';
 
 import Sidebar from './sidebar';
-import Tooltip from './tooltip';
-import InfoBox from './info-box';
-import { EditIcon } from './icons';
+import Tooltip from './components/tooltip';
+import InfoBox from './components/info-box';
+import { EditIcon } from './components/icons';
 import { sanitiseURL } from './helpers';
 
 import CONFIG from './fields-config.json';
