+++ app/src/frontend/components/error-box.tsx
