+++ app/src/frontend/components/info-box.tsx
