+++ app/src/frontend/components/logo.tsx
