+++ app/src/frontend/components/supporter-logos.css
