+++ app/src/frontend/components/supporter-logos.tsx
