+++ app/src/frontend/pages/my-account.tsx
@@ -3,7 +3,7 @@ import { Link, Redirect } from 'react-router-dom';
 import PropTypes from 'prop-types';
 
 import ConfirmationModal from '../confirmation-modal';
-import ErrorBox from '../error-box';
+import ErrorBox from '../components/error-box';
 
 class MyAccountPage extends Component<any, any> { // TODO: add proper types
     static propTypes = { // TODO: generate propTypes from TS
