--- app/src/frontend/welcome.css
@@ -1,23 +1,0 @@
-/**
- * Welcome jumbotron
- */
-.welcome-float {
-    position: absolute;
-    z-index: 10000;
-    top: 0;
-    width: 100%;
-    border-radius: 0;
-    padding: 1.5em 2.5em 2.5em;
-}
-.welcome-float.jumbotron {
-    background: #fff;
-    background-color: rgba(255,255,255,0.95);
-}
-@media (min-width: 768px){
-    .welcome-float {
-        left: 50%;
-        margin-left: -22.5em;
-        width: 45em;
-        top: 1em;
-    }
-}
